api_flask/requirements.txt
Flask==2.0.1
Werkzeug==2.2.3
waitress==2.1.2
google-cloud-firestore==2.1.0
requests==2.26.0
